#### CrossApp 1.1.1 更新内容：
#####1.CAImageView增加序列帧播放功能

#####2.修复图片异步加载失败的bug

#####3.修复webView位置刷新的bug

#####4.修复某些情况安卓真机调试直接崩溃的bug

#####5.修复上一版不能运行iOS模拟器的bug

#####6.其他一些细节的修改

+++ CrossApp/view/CAAlertView.cpp
@@ -350,7 +350,7 @@ void CAAlertView::calcuAlerViewSize()
 	m_fAlertViewHeight += alertViewSpaceHeight;
 
 
-	int btnCount = m_vAllBtn.size();
+	size_t btnCount = m_vAllBtn.size();
 	CCAssert(btnCount,"");
 
 	if (btnCount == 1 || btnCount == 2)
